systemd: add service to selectively disable autorepeat
The xorg-launch-helper does not pass the option to Xorg to disable
autorepeat, but there are several device-specific keys (for Lunchbox)
that should have autorepeat disabled.  These keys include
XF86AudioLowerVolume, XF86AudioRaiseVolume, XF86PowerOff, and XF86Phone.

This commit adds a user-session service file that runs `xset -r KEYCODE1
-r KEYCODE2 ...` to disable autorepeat for these keycodes before e17
starts.  The service file is installed in a separate subpackage that
will be added to Lunchbox images.

Change-Id: I16934eb1ba1c939ce9577ffb02485dea4f77f92d
Signed-off-by: Patrick McCarty <patrick.mccarty@linux.intel.com>